requirements.txt
dashscope
pymysql
langchain
faiss-cpu
sentence-transformers
python-dotenv
